make our own module and exports
- create our own module
- export module to code
